pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "clinical-data-platform"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.111",
    "uvicorn>=0.30",
    "pydantic>=2.7",
    "pandas==2.2.2",
    "scikit-learn==1.5.2",
    "numpy==1.26.4",
    "scipy==1.13.1",
    "pyarrow==16.1.0",
    "markdown==3.6",
    "matplotlib==3.9.2",
    "typer==0.12.3",
    "joblib==1.4.2",
    "Pillow==10.4.0",
    "opentelemetry-sdk==1.25.0",
    "opentelemetry-exporter-otlp-proto-http==1.25.0",
    "opentelemetry-instrumentation-logging==0.45b0",
    "opentelemetry-instrumentation-requests==0.45b0",
    "python-json-logger==2.0.7",
]

[project.optional-dependencies]
dev = [
    "great-expectations>=0.18",
    "mkdocs-material>=9.5",
    "mypy>=1.11",
    "pandera>=0.20",
    "pandas==2.2.2",
    "pytest-cov>=5.0",
    "pytest>=8.2",
    "python-json-logger==2.0.7",
    "ruff>=0.6",
]

[tool.setuptools.packages.find]
where = ["src", "."]
include = ["cp_tox_mini"]

[tool.pytest.ini_options]
addopts = "-q --maxfail=1 --disable-warnings --cov=src --cov-report=xml --cov-report=term"
testpaths = ["tests"]

[tool.coverage.report]
exclude_lines = ["if __name__ == \"__main__\":"]

[tool.ruff]
line-length = 100
src = ["src"]

[tool.mypy]
python_version = "3.10"
strict = false
